pyproject.toml
[tool.poetry]
name = "peppernode_test_task"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.2.3"
djangorestframework = "^3.14.0"
environs = "^9.5.0"
psycopg2 = "^2.9.6"

[tool.poetry.dev-dependencies]
django-stubs = "^4.2.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
